backend/requirements.txt
# Backend - Python Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
google-generativeai>=0.8.0
elevenlabs>=1.0.0
livekit-api>=0.8.0
gTTS>=2.5.0
